refactor: split foundry-cheatcodes defs into a separate crate
