+++ crates/cheatcodes/defs/README.md
@@ -1,0 +1,3 @@
+# foundry-cheatcodes-defs
+
+Minimal crate to provide a common set of cheatcodes definitions.
