+++ crates/cheatcodes/defs/src/items.rs
