+++ crates/cheatcodes/defs/src/vm.rs
