+++ crates/cheatcodes/src/env.rs
@@ -1,7 +1,6 @@
 //! Implementations of [`Environment`](crate::Group::Environment) cheatcodes.
 
-use super::{string, Cheatcode, Result};
-use crate::{Cheatcodes, Vm::*};
+use crate::{string, Cheatcode, Cheatcodes, Result, Vm::*};
 use alloy_primitives::{Address, Bytes, B256, I256, U256};
 use alloy_sol_types::SolValue;
 use std::env;
