+++ crates/cheatcodes/src/lib.rs
@@ -1,99 +1,106 @@
 //! # foundry-cheatcodes
 //!
-//! Foundry cheatcodes definitions and implementations.
+//! Foundry cheatcodes implementations.
 
 #![warn(missing_docs, unreachable_pub, unused_crate_dependencies, rust_2018_idioms)]
 #![allow(elided_lifetimes_in_paths)] // Cheats context uses 3 lifetimes
 
-#[cfg(feature = "impls")]
 #[macro_use]
 extern crate tracing;
 
-// Silence the "unused crate" warning.
-#[cfg(not(feature = "impls"))]
-use alloy_primitives as _;
+use alloy_primitives::Address;
+use foundry_evm_core::backend::DatabaseExt;
+use revm::EVMData;
+use tracing::Level;
 
-pub mod defs;
-pub use defs::{Cheatcode, CheatcodeDef, Vm};
+pub use foundry_cheatcodes_defs::{self as defs, CheatcodeDef, Vm};
 
-#[cfg(feature = "impls")]
-pub mod impls;
-#[cfg(feature = "impls")]
-pub use impls::{Cheatcodes, CheatsConfig};
+#[macro_use]
+mod error;
+pub use error::{Error, ErrorKind, Result};
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::{fs, path::Path};
+mod config;
+pub use config::CheatsConfig;
 
-    const JSON_PATH: &str = concat!(env!("CARGO_MANIFEST_DIR"), "/assets/cheatcodes.json");
-    #[cfg(feature = "schema")]
-    const SCHEMA_PATH: &str = concat!(env!("CARGO_MANIFEST_DIR"), "/assets/cheatcodes.schema.json");
-    const IFACE_PATH: &str = concat!(env!("CARGO_MANIFEST_DIR"), "/../../testdata/cheats/Vm.sol");
+mod inspector;
+pub use inspector::{BroadcastableTransaction, BroadcastableTransactions, Cheatcodes, Context};
 
-    /// Generates the `cheatcodes.json` file contents.
-    fn json_cheatcodes() -> String {
-        serde_json::to_string_pretty(&defs::Cheatcodes::new()).unwrap()
+mod env;
+mod evm;
+mod fs;
+mod json;
+mod script;
+mod string;
+mod test;
+mod utils;
+
+pub use test::expect::ExpectedCallTracker;
+
+/// Cheatcode implementation.
+pub(crate) trait Cheatcode: CheatcodeDef {
+    /// Applies this cheatcode to the given state.
+    ///
+    /// Implement this function if you don't need access to the EVM data.
+    fn apply(&self, state: &mut Cheatcodes) -> Result {
+        let _ = state;
+        unimplemented!("{}", Self::CHEATCODE.func.id)
     }
 
-    /// Generates the [cheatcodes](json_cheatcodes) JSON schema.
-    #[cfg(feature = "schema")]
-    fn json_schema() -> String {
-        serde_json::to_string_pretty(&schemars::schema_for!(defs::Cheatcodes)).unwrap()
+    /// Applies this cheatcode to the given context.
+    ///
+    /// Implement this function if you need access to the EVM data.
+    #[inline(always)]
+    fn apply_full<DB: DatabaseExt>(&self, ccx: &mut CheatsCtxt<DB>) -> Result {
+        self.apply(ccx.state)
     }
 
-    fn sol_iface() -> String {
-        let cheats = defs::Cheatcodes::new().to_string().trim().replace('\n', "\n    ");
-        format!(
-            "\
-// Automatically generated from `foundry-cheatcodes` Vm definitions. Do not modify manually.
-// This interface is just for internal testing purposes. Use `forge-std` instead.
-
-interface Vm {{
-    {cheats}
-}}
-"
-        )
+    #[inline]
+    fn apply_traced<DB: DatabaseExt>(&self, ccx: &mut CheatsCtxt<DB>) -> Result {
+        let span = trace_span(self);
+        let _enter = span.enter();
+        trace_call();
+        let result = self.apply_full(ccx);
+        trace_return(&result);
+        result
     }
+}
 
-    #[test]
-    fn defs_up_to_date() {
-        ensure_file_contents(Path::new(JSON_PATH), &json_cheatcodes());
+// Separate functions to avoid inline and monomorphization bloat.
+fn trace_span<T: Cheatcode>(cheat: &T) -> tracing::Span {
+    if enabled!(Level::TRACE) {
+        trace_span!(target: "cheatcodes", "apply", cheat=?cheat)
+    } else {
+        debug_span!(target: "cheatcodes", "apply", id=%T::CHEATCODE.func.id)
     }
+}
 
-    #[test]
-    #[cfg(feature = "schema")]
-    fn schema_up_to_date() {
-        ensure_file_contents(Path::new(SCHEMA_PATH), &json_schema());
-    }
+fn trace_call() {
+    trace!(target: "cheatcodes", "applying");
+}
 
-    #[test]
-    fn iface_up_to_date() {
-        ensure_file_contents(Path::new(IFACE_PATH), &sol_iface());
-    }
-
-    /// Checks that the `file` has the specified `contents`. If that is not the
-    /// case, updates the file and then fails the test.
-    fn ensure_file_contents(file: &Path, contents: &str) {
-        if let Ok(old_contents) = fs::read_to_string(file) {
-            if normalize_newlines(&old_contents) == normalize_newlines(contents) {
-                // File is already up to date.
-                return
-            }
+fn trace_return(result: &Result) {
+    trace!(
+        target: "cheatcodes",
+        return = match result {
+            Ok(b) => hex::encode(b),
+            Err(e) => e.to_string(),
         }
+    );
+}
 
-        eprintln!("\n\x1b[31;1merror\x1b[0m: {} was not up-to-date, updating\n", file.display());
-        if std::env::var("CI").is_ok() {
-            eprintln!("    NOTE: run `cargo test` locally and commit the updated files\n");
-        }
-        if let Some(parent) = file.parent() {
-            let _ = fs::create_dir_all(parent);
-        }
-        fs::write(file, contents).unwrap();
-        panic!("some file was not up to date and has been updated, simply re-run the tests");
-    }
+/// The cheatcode context, used in [`Cheatcode`].
+pub(crate) struct CheatsCtxt<'a, 'b, 'c, DB: DatabaseExt> {
+    /// The cheatcodes inspector state.
+    pub(crate) state: &'a mut Cheatcodes,
+    /// The EVM data.
+    pub(crate) data: &'b mut EVMData<'c, DB>,
+    /// The original `msg.sender`.
+    pub(crate) caller: Address,
+}
 
-    fn normalize_newlines(s: &str) -> String {
-        s.replace("\r\n", "\n")
+impl<DB: DatabaseExt> CheatsCtxt<'_, '_, '_, DB> {
+    #[inline]
+    pub(crate) fn is_precompile(&self, address: &Address) -> bool {
+        self.data.precompiles.contains(address)
     }
 }
